+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Redmine\Api\AbstractApi::get()` is deprecated, use `\Redmine\Http\HttpClient::request()` instead.
 - `Redmine\Api\AbstractApi::post()` is deprecated, use `\Redmine\Http\HttpClient::request()` instead.
+- `Redmine\Api\AbstractApi::put()` is deprecated, use `\Redmine\Http\HttpClient::request()` instead.
 - The constant `Redmine\Api\Issue::PRIO_LOW` is deprecated.
 - The constant `Redmine\Api\Issue::PRIO_NORMAL` is deprecated.
 - The constant `Redmine\Api\Issue::PRIO_HIGH` is deprecated.